+++ packages/lumx-react/src/components/list/index.ts
@@ -1,4 +1,5 @@
 export * from './List';
 export * from './ListItem';
+export * from './ListItemAction';
 export * from './ListDivider';
 export * from './ListSubheader';